refactor: split stats script into modules
