+++ stats/index.html
@@ -465,6 +465,6 @@
   <script src="utils.js"></script>
   <script src="collections.js"></script>
   <script src="vertical-line.js"></script>
-  <script src="script.js"></script>
+  <script type="module" src="ui.js"></script>
 </body>
 </html>
